+++ tests/testthat/test-new-borders.R
@@ -1,21 +1,21 @@
 
-test_that("bdr objects", {
-  expect_silent(b <- bdr(1, "solid", "red"))
-  expect_s3_class(b, "bdr")
-  expect_true(is_bdr(b))
+test_that("brdr objects", {
+  expect_silent(b <- brdr(1, "solid", "red"))
+  expect_s3_class(b, "brdr")
+  expect_true(is_brdr(b))
   expect_true(is_borderish(b))
   expect_true(is_borderish(1))
   expect_false(is_borderish(-1))
 
-  expect_silent(b <- bdr())
+  expect_silent(b <- brdr())
   expect_equivalent(b$thickness, 0.4)
   expect_equivalent(b$style, "solid")
   expect_equivalent(b$color, NA_character_)
 })
 
 
-test_that("bdr object printing", {
-  expect_silent(b <- bdr(2, "solid", "red"))
+test_that("brdr object printing", {
+  expect_silent(b <- brdr(2, "solid", "red"))
   expect_match(format(b), "2.*solid.*red")
   expect_output(print(b), "2.*solid.*red")
 })
@@ -38,9 +38,9 @@ test_that("left_border(ht) <- number", {
 })
 
 
-test_that("left_border(ht) <- bdr(...)", {
+test_that("left_border(ht) <- brdr(...)", {
   ht <- hux(1:3, 1:3)
-  expect_silent(left_border(ht) <- bdr(1, "double", "red"))
+  expect_silent(left_border(ht) <- brdr(1, "double", "red"))
   expect_equivalent(left_border(ht)[], matrix(1, 3, 2))
   expect_equivalent(left_border_style(ht), matrix("double", 3, 2))
   expect_equivalent(left_border_color(ht), matrix("red", 3, 2))
@@ -55,9 +55,9 @@ test_that("left_border(ht)[i,j] <-number", {
 })
 
 
-test_that("left_border(ht)[i, j] <- bdr(...)", {
+test_that("left_border(ht)[i, j] <- brdr(...)", {
   ht <- hux(1:3, 1:3)
-  expect_silent(left_border(ht)[1:2, 1] <- bdr(1, "double", "red"))
+  expect_silent(left_border(ht)[1:2, 1] <- brdr(1, "double", "red"))
   expect_equivalent(left_border(ht)[1, 1], 1)
 })
 
@@ -73,9 +73,9 @@ test_that("set_left_border(ht, number)", {
 
 
 
-test_that("set_left_border(ht, bdr)", {
+test_that("set_left_border(ht, brdr)", {
   ht <- hux(1:3, 1:3)
-  ht2 <- set_left_border(ht, bdr(1, "double", "red"))
+  ht2 <- set_left_border(ht, brdr(1, "double", "red"))
   expect_equivalent(left_border(ht2)[], matrix(1, 3, 2))
   expect_equivalent(left_border_style(ht2), matrix("double", 3, 2))
   expect_equivalent(left_border_color(ht2), matrix("red", 3, 2))
@@ -99,9 +99,9 @@ test_that("set_left_border(ht, row, col, number)", {
 })
 
 
-test_that("set_left_border(ht, row, col, bdr)", {
+test_that("set_left_border(ht, row, col, brdr)", {
   ht <- hux(1:3, 1:3)
-  ht2 <- set_left_border(ht, 1, 1, bdr(1, "double", "red"))
+  ht2 <- set_left_border(ht, 1, 1, brdr(1, "double", "red"))
   expect_equivalent(left_border(ht2)[], matrix(c(1, 0, 0, 0, 0, 0), 3, 2))
   expect_equivalent(left_border_style(ht2),
         matrix(c("double", rep("solid", 5)), 3, 2))
@@ -114,29 +114,29 @@ test_that("map_left_border", {
   ht2 <- map_left_border(ht, by_ranges(1.5, c(0, 1)))
   expect_equivalent(left_border(ht2)[], matrix(c(0, 1, 1, 0), 2, 2))
 
-  ht3 <- map_left_border(ht, by_ranges(1.5, list(bdr(0), bdr(1))))
+  ht3 <- map_left_border(ht, by_ranges(1.5, list(brdr(0), brdr(1))))
   expect_equivalent(left_border(ht3)[], matrix(c(0, 1, 1, 0), 2, 2))
 
-  ht4 <- map_left_border(ht, by_rows(bdr(0), bdr(1)))
+  ht4 <- map_left_border(ht, by_rows(brdr(0), brdr(1)))
   expect_equivalent(left_border(ht4)[], matrix(c(0, 1, 0, 1), 2, 2))
 
-  ht5 <- map_left_border(ht, by_cols(bdr(0), bdr(1)))
+  ht5 <- map_left_border(ht, by_cols(brdr(0), brdr(1)))
   expect_equivalent(left_border(ht5)[], matrix(c(0, 0, 1, 1), 2, 2))
 
-  # ht8 <- map_left_border(ht, by_cases(. == 2 ~ bdr(1), TRUE ~ bdr(0)))
+  # ht8 <- map_left_border(ht, by_cases(. == 2 ~ brdr(1), TRUE ~ brdr(0)))
   # expect_equivalent(left_border(ht8)[], matrix(c(0, 1, 1, 0), 2, 2))
 
-  ht9 <- map_left_border(ht, by_quantiles(0.5, list(bdr(0), bdr(1))))
+  ht9 <- map_left_border(ht, by_quantiles(0.5, list(brdr(0), brdr(1))))
   expect_equivalent(left_border(ht9)[], matrix(c(0, 1, 1, 0), 2, 2))
 
-  ht10 <- map_left_border(ht, by_equal_groups(2, list(bdr(0), bdr(1))))
+  ht10 <- map_left_border(ht, by_equal_groups(2, list(brdr(0), brdr(1))))
   expect_equivalent(left_border(ht10)[], matrix(c(0, 1, 1, 0), 2, 2))
 
   ht_char <- hux(letters[1:2], letters[2:1])
-  ht6 <- map_left_border(ht_char, by_values("a" = bdr(0), "b" = bdr(1)))
+  ht6 <- map_left_border(ht_char, by_values("a" = brdr(0), "b" = brdr(1)))
   expect_equivalent(left_border(ht6)[], matrix(c(0, 1, 1, 0), 2, 2))
 
-  ht7 <- map_left_border(ht, by_function(function (x) bdr(1)))
+  ht7 <- map_left_border(ht, by_function(function (x) brdr(1)))
   expect_equivalent(left_border(ht7)[], matrix(1, 2, 2))
 })
 
@@ -150,14 +150,14 @@ test_that("map_all_borders", {
   expect_equivalent(right_border(ht2)[2, 2], 0)
   expect_equivalent(left_border(ht2)[1, 1], 0)
 
-  ht3 <- map_all_borders(ht, by_ranges(1.5, list(bdr(0), bdr(1))))
+  ht3 <- map_all_borders(ht, by_ranges(1.5, list(brdr(0), brdr(1))))
   expect_equivalent(left_border(ht3)[2, 1], 1)
   expect_equivalent(bottom_border(ht3)[2, 1], 1)
   expect_equivalent(right_border(ht3)[2, 2], 0)
   expect_equivalent(left_border(ht3)[1, 1], 0)
 
 
-  ht4 <- map_all_borders(ht, 1, 1:2, by_ranges(1.5, list(bdr(0), bdr(1))))
+  ht4 <- map_all_borders(ht, 1, 1:2, by_ranges(1.5, list(brdr(0), brdr(1))))
   expect_equivalent(left_border(ht4)[2, 1], 0)
   expect_equivalent(bottom_border(ht4)[2, 1], 0)
   expect_equivalent(right_border(ht4)[2, 2], 0)
@@ -167,9 +167,9 @@ test_that("map_all_borders", {
 })
 
 
-test_that("set-multiple with bdr", {
+test_that("set-multiple with brdr", {
   ht <- hux(a = 1:2, b = 1:2, add_colnames = FALSE)
-  b <- bdr(2, "dotted", "green")
+  b <- brdr(2, "dotted", "green")
 
   expect_silent(ht2 <- set_all_borders(ht, b))
   expect_equivalent(left_border_color(ht2), matrix("green", 2, 2))